fix: 비디오 duration 정보 추출 및 저장, BigDecimal을 Double로 변경하여 프레임 추출 문제 해결

+++ src/main/java/fakehunters/backend/video/domain/AnalysisResult.java
@@ -2,7 +2,6 @@ package fakehunters.backend.video.domain;
 
 import lombok.Builder;
 import lombok.Getter;
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Getter
@@ -11,7 +10,7 @@ public class AnalysisResult {
     private Long resultId;          // BIGSERIAL
     private Long analysisId;        // FK
     private Boolean isDeepfake;
-    private BigDecimal confidenceScore;
+    private Double confidenceScore;
     private String modelVersion;
     private Long processingTimeMs;
     private String detectedTechniques; // MyBatis Handler를 통해 String 혹은 List로 처리
